Externalize PMMR tests
